[encointerApi] fix type cast

+++ app/lib/service/substrate_api/encointer/encointer_api.dart
@@ -385,7 +385,7 @@ class EncointerApi {
       registration.participantType,
       assignments! as et.Assignment,
       assignmentCount! as et.AssignmentCount,
-      mCount! as int,
+      (mCount! as BigInt).toInt(),
     );
 
     return participantMeetupIndex;
